Refactor json schema validation into separate file/package
