+++ esim/es2/esim/src/test/java/org/ostelco/Es2PlusResourceTest.kt
@@ -12,6 +12,7 @@ import org.ostelco.Es2PlusDownloadOrder
 import org.ostelco.Es2PlusResource
 import org.ostelco.JsonSchema
 import org.ostelco.RestrictedOperationsRequestFilter
+import org.ostelco.jsonValidation.RequestServerReaderWriterInterceptor
 import java.io.*
 import java.util.stream.Collectors
 import javax.ws.rs.WebApplicationException
@@ -20,74 +21,6 @@ import javax.ws.rs.core.MediaType
 import javax.ws.rs.core.Response
 import javax.ws.rs.ext.*
 
-
-class JsonSchemaValidator() {
-    private val schemaRoot = "/es2schemas"
-    private var schemaMap: MutableMap<String, Schema> = mutableMapOf()
-
-
-    private fun loadJsonSchemaResource(name: String) : Schema {
-        val inputStream = this.javaClass.getResourceAsStream("${schemaRoot}/${name}.json")
-        if (inputStream == null) {
-            throw WebApplicationException("Unknown schema map", Response.Status.INTERNAL_SERVER_ERROR)
-        }
-        try {
-            val schema = JSONObject(JSONTokener(inputStream))
-            return org.everit.json.schema.loader.SchemaLoader.load(schema)
-        } catch (e: JSONException) {
-            throw WebApplicationException("Syntax error in schema description", Response.Status.INTERNAL_SERVER_ERROR)
-        }
-    }
-
-
-    private fun getSchema(name: String): Schema {
-        if (!schemaMap.containsKey(name)) {
-            schemaMap[name] = loadJsonSchemaResource(name)
-        }
-        return schemaMap[name]!!
-    }
-
-    public fun validateString(payloadClass: Class<*>, body:String) {
-        val schemaAnnotation = payloadClass.getAnnotation<JsonSchema>(JsonSchema::class.java)
-        if (schemaAnnotation != null) {
-            try {
-                getSchema(schemaAnnotation.schemaKey).validate(JSONObject(body))
-            } catch (t: ValidationException) {
-                throw WebApplicationException(t.errorMessage, Response.Status.BAD_REQUEST)
-            }
-        }
-    }
-}
-
-
-@Provider
-class RequestServerReaderWriterInterceptor : ReaderInterceptor, WriterInterceptor {
-
-    val validator = JsonSchemaValidator()
-
-    @Throws(IOException::class, WebApplicationException::class)
-    override fun aroundReadFrom(ctx: ReaderInterceptorContext): Any {
-        val originalStream = ctx.inputStream
-        val stream = BufferedReader(InputStreamReader(originalStream)).lines()
-        val body: String = stream.collect(Collectors.joining("\n"))
-        ctx.inputStream = ByteArrayInputStream("$body".toByteArray())
-
-        validator.validateString(ctx.type, body)
-
-        return ctx.proceed()
-    }
-
-    @Throws(IOException::class, WebApplicationException::class)
-    override fun aroundWriteTo(ctx: WriterInterceptorContext) {
-
-        // XXX Read docs, figure out how to do this, essentially
-        //     check the  json schema of anything being written, if the
-        //     entity being written has a schema declared in an
-        //     annotation.
-
-        return ctx.proceed()
-    }
-}
 
 
 class ES2PlusResourceTest {
